backend/requirements.txt
# ===========================================
# Tech Roadmap Tracker - Backend Dependencies
# ===========================================

# ----------------- Core -----------------
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1

# ----------------- Database -----------------
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# ----------------- Security -----------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17

# ----------------- Validation & Utils -----------------
email-validator==2.2.0
python-dotenv==1.0.1

# ----------------- Dev & Testing -----------------
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1

# ----------------- Production -----------------
gunicorn==23.0.0
